+++ src/Lib/MrCMS.Web.Admin/Breadcrumbs/FormsBreadcrumb.cs
@@ -4,7 +4,7 @@ namespace MrCMS.Web.Admin.Breadcrumbs
 {
     public class FormsBreadcrumb : Breadcrumb
     {
-        public override int Order => 3;
+        public override decimal Order => 3;
         public override string Controller => "Form";
         public override string Action => "Index";
         public override bool IsNav => true;
