+++ src/Lib/MrCMS.Web.Admin/Breadcrumbs/MediaBreadcrumb.cs
@@ -4,7 +4,7 @@ namespace MrCMS.Web.Admin.Breadcrumbs
 {
     public class MediaBreadcrumb : Breadcrumb
     {
-        public override int Order => 1;
+        public override decimal Order => 1;
         public override string Controller => "MediaCategory";
         public override string Action => "Index";
         public override bool IsNav => true;
