+++ src/Lib/MrCMS.Web.Admin/Breadcrumbs/System/AboutBreadcrumb.cs
@@ -4,7 +4,7 @@ namespace MrCMS.Web.Admin.Breadcrumbs.System
 {
     public class AboutBreadcrumb : Breadcrumb<SystemBreadcrumb>
     {
-        public override int Order => 15;
+        public override decimal Order => 15;
         public override string Controller => "About";
         public override string Action => "Index";
         public override bool IsNav => true;
