+++ src/Lib/MrCMS.Web.Admin/Breadcrumbs/System/BatchesBreadcrumb.cs
@@ -4,7 +4,7 @@ namespace MrCMS.Web.Admin.Breadcrumbs.System
 {
     public class BatchesBreadcrumb : Breadcrumb<SystemBreadcrumb>
     {
-        public override int Order => 9;
+        public override decimal Order => 9;
         public override string Controller => "Batch";
         public override string Action => "Index";
         public override bool IsNav => true;
