+++ src/Lib/MrCMS.Web.Admin/Breadcrumbs/System/ClearCachesBreadcrumb.cs
@@ -4,7 +4,7 @@ namespace MrCMS.Web.Admin.Breadcrumbs.System
 {
     public class ClearCachesBreadcrumb : Breadcrumb<SystemBreadcrumb>
     {
-        public override int Order => 14;
+        public override decimal Order => 14;
         public override string Controller => "ClearCaches";
         public override string Action => "Index";
         public override bool IsNav => true;
