+++ src/Lib/MrCMS.Web.Admin/Breadcrumbs/System/ImportExportDocumentsBreadcrumb.cs
@@ -4,7 +4,7 @@ namespace MrCMS.Web.Admin.Breadcrumbs.System
 {
     public class ImportExportWebpagesBreadcrumb : Breadcrumb<SystemBreadcrumb>
     {
-        public override int Order => 2;
+        public override decimal Order => 2;
         public override string Name => "Import/Export Webpages";
         public override string Controller => "ImportExport";
         public override string Action => "Documents";
