+++ src/Lib/MrCMS.Web.Admin/Breadcrumbs/System/LogsBreadcrumb.cs
@@ -4,7 +4,7 @@ namespace MrCMS.Web.Admin.Breadcrumbs.System
 {
     public class LogsBreadcrumb : Breadcrumb<SystemBreadcrumb>
     {
-        public override int Order => 8;
+        public override decimal Order => 8;
         public override string Controller => "Log";
         public override string Action => "Index";
         public override bool IsNav => true;
