+++ src/Lib/MrCMS.Web.Admin/Breadcrumbs/System/MessageQueueBreadcrumb.cs
@@ -4,7 +4,7 @@ namespace MrCMS.Web.Admin.Breadcrumbs.System
 {
     public class MessageQueueBreadcrumb : Breadcrumb<SystemBreadcrumb>
     {
-        public override int Order => 12;
+        public override decimal Order => 12;
         public override string Controller => "MessageQueue";
         public override string Action => "Index";
         public override bool IsNav => true;
