+++ src/Lib/MrCMS.Web.Admin/Breadcrumbs/System/MessageTemplatesBreadcrumb.cs
@@ -4,7 +4,7 @@ namespace MrCMS.Web.Admin.Breadcrumbs.System
 {
     public class MessageTemplatesBreadcrumb : Breadcrumb<SystemBreadcrumb>
     {
-        public override int Order => 3;
+        public override decimal Order => 3;
         public override string Controller => "MessageTemplate";
         public override string Action => "Index";
         public override bool IsNav => true;
