+++ src/Lib/MrCMS.Web.Admin/Breadcrumbs/System/NotificationsBreadcrumb.cs
@@ -4,7 +4,7 @@ namespace MrCMS.Web.Admin.Breadcrumbs.System
 {
     public class NotificationsBreadcrumb : Breadcrumb<SystemBreadcrumb>
     {
-        public override int Order => 13;
+        public override decimal Order => 13;
         public override string Controller => "Notification";
         public override string Action => "Index";
         public override bool IsNav => true;
