+++ src/Lib/MrCMS.Web.Admin/Breadcrumbs/System/PageDefaultsBreadcrumb.cs
@@ -4,7 +4,7 @@ namespace MrCMS.Web.Admin.Breadcrumbs.System
 {
     public class PageDefaultsBreadcrumb : Breadcrumb<SystemBreadcrumb>
     {
-        public override int Order => 5;
+        public override decimal Order => 5;
         public override string Controller => "PageDefaults";
         public override string Action => "Index";
         public override bool IsNav => true;
