+++ src/Lib/MrCMS.Web.Admin/Breadcrumbs/System/PageTemplatesBreadcrumb.cs
@@ -4,7 +4,7 @@ namespace MrCMS.Web.Admin.Breadcrumbs.System
 {
     public class PageTemplatesBreadcrumb : Breadcrumb<SystemBreadcrumb>
     {
-        public override int Order => 4;
+        public override decimal Order => 4;
         public override string Controller => "PageTemplate";
         public override string Action => "Index";
         public override bool IsNav => true;
