+++ src/Lib/MrCMS.Web.Admin/Breadcrumbs/System/Settings/SystemSettingsBreadcrumb.cs
@@ -8,6 +8,6 @@ namespace MrCMS.Web.Admin.Breadcrumbs.System.Settings
         public override string Action => "Index";
 
         public override bool IsNav => true;
-        public override int Order => 1;
+        public override decimal Order => 1;
     }
 }
