+++ src/Lib/MrCMS.Web.Admin/Breadcrumbs/System/SitesBreadcrumb.cs
@@ -4,7 +4,7 @@ namespace MrCMS.Web.Admin.Breadcrumbs.System
 {
     public class SitesBreadcrumb : Breadcrumb<SystemBreadcrumb>
     {
-        public override int Order => 6;
+        public override decimal Order => 6;
         public override string Controller => "Sites";
         public override string Action => "Index";
         public override bool IsNav => true;
