+++ src/Lib/MrCMS.Web.Admin/Breadcrumbs/System/TasksBreadcrumb.cs
@@ -4,7 +4,7 @@ namespace MrCMS.Web.Admin.Breadcrumbs.System
 {
     public class TasksBreadcrumb : Breadcrumb<SystemBreadcrumb>
     {
-        public override int Order => 10;
+        public override decimal Order => 10;
         public override string Controller => "Task";
         public override string Action => "Index";
         public override bool IsNav => true;
